backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
PyJWT[crypto]>=2.4.0
bcrypt>=4.0.0
python-multipart==0.0.9
pydantic==2.6.1
httpx==0.27.0
aiofiles==23.2.1
ldap3==2.9.1
slowapi==0.1.9
